requirements.txt
certifi==2019.6.16
chardet==3.0.4
DateTime==4.3
httplib2==0.13.0
idna==2.8
pprint==0.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
PySimpleGUI==4.0.0
python-dotenv==0.10.3
pytz==2019.1
requests==2.22.0
rsa==4.0
six==1.12.0
urllib3==1.25.3
zope.interface==4.6.0